Issue #80 - De-unify dom/xhr
